prevent that big fixtures.tar.gz files is included in vendor copies
When the package is vendored as a go module dependency in another
project the fixtures.tar.gz becomes part of the copy.

The file unnecessarily increases the size of vendor directories.
It is 15MB big and only used in testcases.
Testcases are not included in vendor copies.

Move the file to a testdata/ directory, to prevent that it's included in
vendor copies.