modules: fix: version 2.0.0 is not installable via go-tooling
A git tag and github release was created for a version 2.0.0 but the
version was not installable via go tools and could not be added as
dependency to a go.mod file.

For Major version >1 the version must be appended to the import path.
This was missing. (see
https://github.com/golang/go/wiki/Modules#releasing-modules-v2-or-higher
for more information)

Append /v2 to the import path, change all references accordingly.
The retract directives for minor versions of v1 are removed. They can
only be defined in the go.mod file for the same major version.